requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
psycopg2-binary>=2.9.1
aiofiles>=0.7.0
mutagen>=1.45.1
python-dotenv>=0.19.0
alembic>=1.7.1 


pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
factory-boy==3.3.0
pytest-xdist==3.3.1
pytest-watch==4.2.0
pytest-cov==4.1.0
